revert: feat: add custom cwood social icon
This reverts commit 6280036e22fc0ed97ba3dfe5c07384ed2a662792.

These feels unnecessary given we have a link to the site on the
homepage, and we now redir the base domains to the official site.

In addition, it just doesn't look as good. For example, viewers that
use the Dark Reader browser extension will see the custom icon in a
different color than the rest.